+++ DoubleFile/DoubleFile/UserControls/MVVM_Clones/UC_ClonesVM_Impl.cs
@@ -62,7 +62,7 @@ namespace DoubleFile
             var folderDetail = LocalTV.TreeSelect_FolderDetail;
 
             if (null != folderDetail)
-                TreeSelect_FolderDetailUpdated(Tuple.Create(folderDetail, 0));
+                TreeSelect_FolderDetailUpdated(Tuple.Create(folderDetail, /* UI initiator */ 0m));
 
             Icmd_GoTo = new RelayCommand(GoTo, () => null != _selectedItem);
 
@@ -89,7 +89,7 @@ namespace DoubleFile
             Util.LocalDispose(_lsDisposable);
         }
 
-        void TreeSelect_FolderDetailUpdated(Tuple<TreeSelect.FolderDetailUpdated, int> initiatorTuple)
+        void TreeSelect_FolderDetailUpdated(Tuple<TreeSelect.FolderDetailUpdated, decimal> initiatorTuple)
         {
             var folderDetail = initiatorTuple.Item1;
